@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.3'
 
     defaultConfig {
         applicationId "wsy.org.mytestapplication"
         minSdkVersion 14
-        targetSdkVersion 27
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
     }
@@ -17,19 +17,31 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
+    applicationVariants.all { variant ->
+        Object object = "Object"
+        int result = testMethod(object)
+        print(result)
+    }
+}
+
+static int testMethod(String arg) {
+    return 1;
+}
+
+static int testMethod(Object arg) {
+    return 2;
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation project(':navtablayout')
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:support-v4:26.0.2'
-    compile 'com.android.support:recyclerview-v7:26.0.2'
-
-    compile 'io.reactivex:rxjava:1.0.14'
-    compile 'com.github.bumptech.glide:glide:3.7.0'
-    compile 'com.github.bumptech.glide:glide:3.7.0'
-    compile 'com.android.support:design:27.1.1'
-    compile 'com.android.support:preference-v14:26.0.2'
-    compile 'de.greenrobot:eventbus:3.0.0-beta1'
+    testImplementation 'junit:junit:4.12'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'io.reactivex:rxjava:1.0.14'
+    implementation 'com.github.bumptech.glide:glide:4.10.0'
+    implementation 'com.google.android.material:material:1.1.0'
+    implementation 'androidx.legacy:legacy-preference-v14:1.0.0'
+    implementation 'de.greenrobot:eventbus:3.0.0-beta1'
+    implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.0.0'
 }
