@@ -2,7 +2,7 @@ package wsy.org.mytestapplication.activity;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import wsy.org.mytestapplication.R;
 
